[requests] Upgrade to `2.6.2`
Compared to `2.6.0`, `2.6.1` updates the way the modules vendored
with `requests` in `requests.packages.*` are imported. In particular
this allows `requests` to import `requests.packages.urllib3` and not
try to import the `urllib3` module directly.

We upgrade to `2.6.2` directly because it fixes a regression introduced
by `2.6.1`.

See https://github.com/kennethreitz/requests/blob/v2.6.2/HISTORY.rst

Why the urllib3 direct import was problematic:
Version `1.17` of `urllib3` (used by `python-etcd`) introduced changes
to its deps and the way it loads its deps. Unfortunately these changes
incurred a lot of new imports from the `requests` 2.6.0 module, which
lead to ~5 to 10MB of RSS overhead from the python process.