[urllib3] Pin to `1.16`
Version `1.17` of `urllib3` (dep of `python-etcd`) introduces changes to
its deps and the way it loads its deps for its pyopenssl contrib module

Unfortunately these changes cause a lot of new imports (`OpenSSL`,
`idna`, `cryptography`) which lead to increases of the RSS of the python
processes (about 5MB per process).

AFAIK we do not need the changes brought by the `1.17` upgrade in
`python-etcd`, so let's pin `urllib3` to `1.16`.