[guidata] new source
Guidata was hosted on Google Code, which is now dead for real.
This commit switches the source to Github. The md5 changes because the
previous zip was slightly compressed (and was not shipping the tests for
instance), whereas this one contains the entire source code (because
it's automcatically generated by Github).