[gohai] use glide for dependency management
We completely stop using go get. We do it the proper way.
The binary generated is now named 'datadog-gohai', so the build is also
adapted.

This removes always_build as it is not needed: by defining the source
as `github: "DataDog/gohai"` omnibus rebuilds it when changes are pushed
to the `last-stable` branch.