[gohai] Fix build with old version of go (1.3)
For some reason, when the deps of gohai are not yet present
in the gopath, `go get`ting gohai fails by getting the latest version
of gopsutil, which is only compatible with go >= 1.8.0. This happens
even with the `go get -d` option, which is supposed to only download
the packages.

As a workaround, we `go get` gohai only once the deps are already
present and pinned to supported versions.

The whole software def is quite ugly, that'll improve once we can
start using `glide` (which is blocked on the upgrade of go, which
itself waits on the upgrade of the centos build image to centos 6)