[jmxfetch] Move bintray enpdoints to sonatype for Agent v5
Since bintray is sutting down, this change ensures that we get our
artifacts from Sonatype instead.

Note: Since artifacts for Agent 5 were not on Sonatype previously, a
manual publish was done but for those specific versions the filename
was mangled by Sonatype so there are overrides for the URL endpoint to
ensure that we are retrieving the proper file.